requirements.txt
aiodns==3.1.1
aiofiles==23.2.1
aiohttp==3.8.6
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
Brotli==1.1.0
bs4==0.0.1
camelot-py==0.11.0
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.1
click==8.1.7
cryptography==41.0.5
cssselect==1.2.0
distro==1.8.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
fake-useragent==1.3.0
frozenlist==1.4.0
ghostscript==0.7
h11==0.14.0
html5lib==1.1
idna==3.4
importlib-metadata==6.8.0
JPype1==1.4.1
lxml==4.9.3
mechanize==0.4.8
multidict==6.0.4
numpy==1.26.1
opencv-python==4.8.1.78
openpyxl==3.1.2
outcome==1.3.0
packaging==23.2
pandas==2.1.1
parse==1.19.1
pdfminer.six==20221105
pycares==4.4.0
pycparser==2.21
pyee==8.2.2
pypdf==3.16.4
pyppeteer==1.0.2
pyquery==2.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
requests-html==0.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
tabula-py==2.8.2
tabulate==0.9.0
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.11.1
tzdata==2023.3
urllib3==1.26.18
w3lib==2.1.2
webdriver-manager==4.0.1
webencodings==0.5.1
websockets==10.4
wsproto==1.2.0
yarl==1.9.2
zipp==3.17.0
